+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Release Notes for Multi Translator
 
-## Unreleased
+## 2.23.0 - 2025-02-05
+
+### Added
 
 - Add support for Link (native) fields
+
+### Fixed
+
 - map 'no' to 'nb' for DeepL
-- fix translating empty linkit fields (fixes #98)
+- fix translating empty linkit fields
 
 ## 2.22.2 - 2025-01-28